\name{flowerSize}
\docType{data}
\alias{flowerSize}
\title{Flower size of 25 Euphorbiaceae species}
\description{ Names, flower diameters (mm) and log-transformed diameter (mm) of 25 plant species.}
\usage{data(flowerSize)}
\format{A data frame with 25 rows and 3 columns.}
\references{
Davis, C.C., Latvis, M., Nickrent, D.L., Wurdack, K.J. and Baum, D.A. 2007. "Floral gigantism in Rafflesiaceae". Science \bold{315}:1812.
}
\keyword{datasets}

