\name{flowerTree}
\docType{data}
\alias{flowerTree}
\title{Phylogenetic tree of 25 Euphorbiaceae species}
\description{A phylogenetic tree with 25 tips and 24 internal nodes.}
\usage{data(flowerTree)}
\format{A data frame of class phylo.}
\references{
Davis, C.C., Latvis, M., Nickrent, D.L., Wurdack, K.J. and Baum, D.A. 2007. "Floral gigantism in Rafflesiaceae". Science \bold{315}:1812.
}
\keyword{datasets}