\name{guidetree}
\docType{data}
\alias{guidetree}
\title{Binary population tree within Arabidopsis thaliana}
\description{Binary population tree for
29 A. thaliana accessions and A. lyrata,
obtained from chromosome 4 using MDL to delimit loci,
BUCKy to estimate quartet concordance factors (CFs) and
Quartet Max Cut to estimate the tree topology.}

\usage{data(guidetree)}

\format{tree object of class phylo. Branch lengths are in coalescent units.}

\references{
Stenz, Noah W. M., Bret Larget, David A. Baum and Cécile Ané (2015).
Exploring tree-like and non-tree-like patterns using genome sequences:
An example using the inbreeding plant species \emph{Arabidopsis thaliana} (L.) Heynh.
Systematic Biology, \bold{64}(5):809-823.

TICR pipeline: github.com/nstenz/TICR
}
\keyword{datasets}