\name{OUshifts-methods}
\alias{plot.OUshifts}

\title{Methods for class 'OUshifts'.}

\description{These are method functions for class 'OUshifts'.}

\usage{
\method{plot}{OUshifts}(x, show.data = TRUE, digits=3, ...)
}

\arguments{      
  \item{x}{an object of class \code{"OUshifts"}.}
  \item{show.data}{if TRUE, visualizes a bar plot of the data side-by-side with the phylogeny.}
  \item{digits}{number of digits to show in the bar plot.}
  \item{\dots}{further arguments passed to plot.phylo to plot the tree.}
}

\author{Lam Si Tung Ho, Kelsey Tarpinian}

\seealso{
\code{\link{OUshifts}}
}




