\encoding{latin1}
\name{phylolm-package}
\alias{phylolm-package}
\docType{package}
\title{Phylogenetic Linear Regression}
\description{The phylolm package provides a function for fitting phylogenetic linear regression models. The likelihood is calculated with an algorithm that is linear in the number of tips in the tree. The package also provides a function for  simulating continuous traits along the tree.}
\details{
\tabular{ll}{
Package: \tab phylolm\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2012-09-21\cr
License: \tab GPL (>= 2)\cr
}
}
\author{Lam Si Tung Ho, Cécile Ané

Maintainer: Lam Si Tung Ho <lamho@stat.wisc.edu>
}
