\encoding{latin1}
\name{phylolm-package}
\alias{phylolm-package}
\docType{package}
\title{Phylogenetic Linear Regression}
\description{The package provides functions for fitting phylogenetic linear models and phylogenetic generalized linear models. The computation uses an algorithm that is linear in the number of tips in the tree. The package also provides functions for simulating continuous and binary traits along the tree.
Other tools include functions to test the adequacy of a population tree.}
\details{
\tabular{ll}{
Package: \tab phylolm\cr
Type: \tab Package\cr
Version: \tab 2.7.0\cr
Date: \tab 2026-02-23\cr
License: \tab GPL (>= 2)\cr
}
}
\author{Lam Si Tung Ho, Cecile Ane, Robert Lachlan, Kelsey Tarpinia, Rachel Feldman, Qing Yu, Wouter van der Bijl, Joan Maspons, Rutger Vos, Paul Bastide, Ana Marcia Barbosa

Maintainer: Lam Si Tung Ho <lamho86@gmail.com>
}
